Safari-611.1.10/LayoutTests/crypto/subtle/ecdh-import-key-derive-bits-null-length-expected.txt

Test ECDH deriveBits operation with imported base key and null length

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS bytesToHexString(derivedKey) is expectedDerivedKey
PASS successfullyParsed is true

TEST COMPLETE
